ui & stacking
Updated healing bomb to make ui better & items dont stack if they aren't the same infusion
