+++ module.json
@@ -2,7 +2,7 @@
     "id": "pf2e-alchemist-remaster-ducttape-fork",
     "title": "PF2e Alchemist Remaster Duct Tape Fork",
     "description": "A module to slap together some functionality until the PF2e system is updated to support the Alchmist remaster.",
-    "version": "2.8.6",
+    "version": "2.8.7",
     "compatibility": {
         "minimum": "12",
         "verified": "12"
@@ -13,6 +13,7 @@
     "esmodules": [
         "scripts/settings.js",
         "scripts/QuickAlchemy.js",
+        "scripts/HealingBomb.js",
         "scripts/AlchemistFeats.js",
         "scripts/FormulaSearch.js",
         "scripts/LevelUp.js",
