+++ packs/alchemist-duct-tape-items/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000184
+MANIFEST-000193
