+++ packs/alchemist-duct-tape-macros/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000253
+MANIFEST-000263
