+++ packs/alchemist-duct-tape-macros/LOG.old
@@ -1,3 +1,60 @@
-2025/04/11-20:46:01.313 3008 Recovering log #249
-2025/04/11-20:46:01.324 3008 Delete type=0 #249
-2025/04/11-20:46:01.324 3008 Delete type=3 #247
+2025/04/12-09:02:56.042 5300 Recovering log #254
+2025/04/12-09:02:56.042 5300 Recovering log #259
+2025/04/12-09:02:56.072 5300 Delete type=2 #5
+2025/04/12-09:02:56.072 5300 Delete type=0 #6
+2025/04/12-09:02:56.072 5300 Delete type=0 #28
+2025/04/12-09:02:56.072 5300 Delete type=2 #30
+2025/04/12-09:02:56.072 5300 Delete type=0 #40
+2025/04/12-09:02:56.072 5300 Delete type=0 #62
+2025/04/12-09:02:56.073 5300 Delete type=0 #64
+2025/04/12-09:02:56.073 5300 Delete type=0 #72
+2025/04/12-09:02:56.073 5300 Delete type=0 #74
+2025/04/12-09:02:56.073 5300 Delete type=0 #76
+2025/04/12-09:02:56.073 5300 Delete type=0 #78
+2025/04/12-09:02:56.073 5300 Delete type=0 #80
+2025/04/12-09:02:56.073 5300 Delete type=0 #92
+2025/04/12-09:02:56.073 5300 Delete type=0 #100
+2025/04/12-09:02:56.074 5300 Delete type=0 #106
+2025/04/12-09:02:56.074 5300 Delete type=0 #108
+2025/04/12-09:02:56.074 5300 Delete type=0 #111
+2025/04/12-09:02:56.074 5300 Delete type=0 #127
+2025/04/12-09:02:56.074 5300 Delete type=0 #158
+2025/04/12-09:02:56.074 5300 Delete type=0 #161
+2025/04/12-09:02:56.074 5300 Delete type=0 #166
+2025/04/12-09:02:56.075 5300 Delete type=0 #184
+2025/04/12-09:02:56.075 5300 Delete type=0 #186
+2025/04/12-09:02:56.075 5300 Delete type=0 #193
+2025/04/12-09:02:56.075 5300 Delete type=0 #196
+2025/04/12-09:02:56.075 5300 Delete type=2 #198
+2025/04/12-09:02:56.075 5300 Delete type=0 #200
+2025/04/12-09:02:56.075 5300 Delete type=0 #254
+2025/04/12-09:02:56.075 5300 Delete type=0 #259
+2025/04/12-09:02:56.075 5300 Delete type=3 #4
+2025/04/12-09:02:56.076 5300 Delete type=3 #26
+2025/04/12-09:02:56.076 5300 Delete type=3 #39
+2025/04/12-09:02:56.076 5300 Delete type=3 #61
+2025/04/12-09:02:56.076 5300 Delete type=3 #63
+2025/04/12-09:02:56.077 5300 Delete type=3 #71
+2025/04/12-09:02:56.077 5300 Delete type=3 #73
+2025/04/12-09:02:56.077 5300 Delete type=3 #75
+2025/04/12-09:02:56.077 5300 Delete type=3 #77
+2025/04/12-09:02:56.077 5300 Delete type=3 #79
+2025/04/12-09:02:56.078 5300 Delete type=3 #91
+2025/04/12-09:02:56.078 5300 Delete type=3 #99
+2025/04/12-09:02:56.078 5300 Delete type=3 #105
+2025/04/12-09:02:56.078 5300 Delete type=3 #107
+2025/04/12-09:02:56.078 5300 Delete type=3 #109
+2025/04/12-09:02:56.078 5300 Delete type=3 #125
+2025/04/12-09:02:56.078 5300 Delete type=3 #157
+2025/04/12-09:02:56.079 5300 Delete type=3 #159
+2025/04/12-09:02:56.079 5300 Delete type=3 #165
+2025/04/12-09:02:56.079 5300 Delete type=3 #183
+2025/04/12-09:02:56.079 5300 Delete type=3 #185
+2025/04/12-09:02:56.079 5300 Delete type=3 #192
+2025/04/12-09:02:56.079 5300 Delete type=3 #194
+2025/04/12-09:02:56.079 5300 Delete type=3 #199
+2025/04/12-09:02:56.079 5300 Delete type=3 #253
+2025/04/12-09:40:43.510 4238 Level-0 table #262: started
+2025/04/12-09:40:43.510 4238 Level-0 table #262: 0 bytes OK
+2025/04/12-09:40:43.514 4238 Delete type=0 #260
+2025/04/12-09:40:43.538 4238 Manual compaction at level-0 from '!macros!A8Fd0G580CovRtDq' @ 72057594037927935 : 1 .. '!macros!ZTOUwRh7BGixOtL9' @ 0 : 0; will stop at (end)
